+++ src/main/java/com/example/oop6/models/shapes/ShapeGroup.java
@@ -3,14 +3,13 @@ package com.example.oop6.models.shapes;
 import com.example.oop6.models.shapes.funcs.ShapeAction;
 import com.example.oop6.utils.Container;
 import javafx.scene.canvas.GraphicsContext;
-import javafx.scene.paint.Color;
 
 public class ShapeGroup extends Shape {
     private final Container<Shape> shapes;
-    private int leftShape = 0;
-    private int rightShape = 0;
-    private int upShape = 0;
-    private int downShape = 0;
+    private double leftShape = 0;
+    private double rightShape = 0;
+    private double upShape = 0;
+    private double downShape = 0;
 
     public ShapeGroup() {
         super(0, 0);
@@ -38,8 +37,8 @@ public class ShapeGroup extends Shape {
         if (shape.position.getY() < upShape) upShape = shape.position.getY() - shape.getMinHeight() / 2;
         width = maxX - minX;
         height = maxY - minY;
-        int oldX = position.getX();
-        int oldY = position.getY();
+        double oldX = position.getX();
+        double oldY = position.getY();
         position.setX((int) (minX + width / 2));
         position.setY((int) (minY + height / 2));
         //нормализация координат относительно центра группы
@@ -75,7 +74,7 @@ public class ShapeGroup extends Shape {
     }
 
     @Override
-    public boolean inShapeArea(int x, int y) {
+    public boolean inShapeArea(double x, double y) {
         for (Shape shape : shapes) {
             shape.getPosition().changePosition(position.getX(), position.getY());
             boolean res = shape.inShapeArea(x, y);
@@ -102,24 +101,24 @@ public class ShapeGroup extends Shape {
 
     //переименовать
     @Override
-    public boolean entersByWidth(int width) {
+    public boolean entersByWidth(double width) {
         boolean res = super.entersByWidth(width);
         return res && this.width >= rightShape - leftShape;
     }
 
     @Override
-    public boolean entersByHeight(int height) {
+    public boolean entersByHeight(double height) {
         boolean res = super.entersByHeight(height);
         return res && this.height >= downShape - upShape;
     }
 
     @Override
-    public int getMinWidth() {
+    public double getMinWidth() {
         return rightShape - leftShape;
     }
 
     @Override
-    public int getMinHeight() {
+    public double getMinHeight() {
         return downShape - upShape;
     }
 }
